requirements.txt
uvicorn
fastapi
gunicorn
motor
pymongo
pymongo[srv]
python-dotenv
jinja2